Move compiler to src
