+++ src/main.rs
@@ -4,12 +4,13 @@ mod parser;
 mod checker;
 mod codegen;
 mod assembler;
+mod compiler;
 mod flags;
 mod new;
 
 use std::time::Instant;
 
-use crate::new::compiler::main as other_main;
+use crate::compiler::main as other_main;
 fn main() {
     // if let Err(e) = compile() {
     //     println!("{}", e);
@@ -21,7 +22,7 @@ fn main() {
 mod tests {
     use lazy_static::lazy_static;
 
-    use crate::{new::compiler::Compiler, codegen::ERR_STR};
+    use crate::{compiler::Compiler, codegen::ERR_STR};
 
     const ALWAYS_FAILS: &str = "This is a String we defined to make sure that a test always fails. This is expected.";
 
@@ -73,7 +74,7 @@ mod tests {
     }
 
     mod syntax_tests {
-        use crate::new::{compiler::Compiler};
+        use crate::compiler::Compiler;
         use crate::codegen::ERR_STR;
 
         macro_rules! generate_failing_test {
@@ -105,7 +106,7 @@ mod tests {
     }
 
     mod semantic_tests {
-        use crate::{new::compiler::Compiler, codegen::{ERR_STR}};
+        use crate::{compiler::Compiler, codegen::{ERR_STR}};
         use crate::tests::ALWAYS_FAILS;
 
         macro_rules! generate_failing_test {
