+++ src/new/mod.rs
@@ -1,3 +1,2 @@
 pub mod nodes;
 pub mod instr;
-pub mod compiler;
